refactor: merge tally.go with ballot.go
